# swg_api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DADS API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://portal.dratdoorstep.com/api"
        },
        {
            "url": "https://deviboon.dratdoorstep.com/api"
        },
        {
            "url": "http://localhost/DADS_Web/api"
        }
    ],
    "paths": {
...
        "/PromoCode/get": {
            "get": {
                "tags": [
                    "Promo Code"
                ],
                "summary": "get Promo Code Details",
                "description": "",
                "operationId": "getPromoCode",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid/unauthorized token or token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dads_auth": []
                    }
                ]
            }
        },
        "/PromoCode/add": {
            "post": {
                "tags": [
                    "Promo Code"
                ],
                "summary": "Validate Promo Code",
                "description": "",
                "operationId": "postPromoCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromoCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to insert Address profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid/unauthorized token or token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Parameters missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dads_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PromoCodeRequest": {
                "required": [
                    "promo_code",
                    "cart_id"
                ],
                "properties": {
                    "promo_code": {
                        "type": "string"
                    },
                    "cart_id": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "type": "object"
            },
...
            "Api_Response": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "dads_auth": {
                "type": "http",
                "name": "dads_auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}